fix: Abort tasks on retry (#951)
# Motivation

While handling a processing task from a crashed pod, if there is no way
to complete the task on behalf of the crashed pod, results created by
the task were not aborted, and OwnerTaskId of ExpectedOutputKeys were
not reset.

# Description

Restore the OwnerTaskId and abort all created tasks and results before
retrying the task.
The implementation of AbortTasksAndResults has also been refactored to
support aborting specific results in addition to tasks.
The reason messages for the abortion have also been adjusted to better
reflect the actual reason.

# Impact

In case of retry after crash, there should be no more cases of pending
tasks and results.
Error recovery will be a bit slower because of the additional queries.