fix: Idem-potent retry (#936)
# Motivation

If 2 pods retry the same task, current code makes it that only of them
will actually retry the task. In some cases, the pod that should be
responsible for the retry crashes, and the retry is never submitted.

# Description

Avoid the election of a pod responsible for the retry, and let both of
them submit the retry task.
Refactor task ending and retry to mutualize code.

# Testing

A unit test has been added that start multiple tasks retrying the same
task concurrently, for all task status possible, even Retried. The test
check that the retried task has successfully been created and submitted
to the queue.

# Impact

This enables error recovery when an agent crashes while creating a retry
task..