refactor: Expose attributes parameter when creating SQS (#955)
# Motivation

SQS supports many options when creating SQS queues that ArmoniK does not
expose.

# Description

Expose the dictionary config to be passed when creating SQS queues.

# Testing

Deploying on AWS with attribute to set max message size to 1024 works as
expected.

# Impact

Enable more customization when creating the SQS, be it SQS policy,
retention, or KMS configuration.

# Checklist

- [x] My code adheres to the coding and style guidelines of the project.
- [x] I have performed a self-review of my code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have made corresponding changes to the documentation.
- [x] I have thoroughly tested my modifications and added tests when
necessary.
- [x] Tests pass locally and in the CI.
- [x] I have assessed the performance impact of my modifications.