refactor: MongoDB BaseTable and read preference (#1031)
# Motivation

Monitoring RPCs (like list and count) are heavy on the database, but do
not target secondary servers.
This imposes a high load on the primary server, while those RPCs do not
require the consistency offered by the primary.

# Description

Refactor MongoDB Adapter to mutualize code between tables, and implement
read-only optimizations to target read replicas for monitoring services.

# Testing

Deploy ArmoniK with 3 MongoDB replicas.
Populate the database with many tasks (> 100k), and go to the list of
session page on the GUI.
Configure the auto refresh to 0.1s.

With main, only the primary has some loads. with this PR, only
secondaries have load.
If there is a single replicas, it still works and the load goes
(obviously) to the primary.

# Impact

Read RPCs are configured to target read replicas. It might be possible
to have inconsistencies with writes that are done on the primary if the
read RPC is performed too quickly after the write. To limit this impact,
read preference has been enabled only on Count RPCs.

# Additional Information

More refactor can be done to integrate SessionProvider and
CollectionProvider within BaseTable.