fix: tasks and results creation should be atomic (#1063)
We observed cases of tasks that were not properly submitted due to
interruptions or errors during submission. This PR aims to make
submission related RPCs atomic so that, when there is an issue, no tasks
or results are left in an inconsistent state.

A try/catch block was added around critical parts of the code where the
tasks/resutls are deleted from the database/object storage upon
exception.

Unit tests were implemented to check that tasks and results are properly
deleted.

- Tasks that are in Pending state due to errors during submission should
be deleted and not appear in monitoring anymore.
- Issues during result creation should not leave partial results created

- Rollback after task submission by another task is not properly
implemented because owner id in the result is changed without keeping
the old value. It requires a change in the data scheme to keep the old
value and put it back on if needed. This case should be treated at some
point as it is needed to properly rollback sub tasks when task is
reacquired after crash during postprocessing.
- This correction is not completely foolproof as it assumes the service
to be able to execute the catch clause. It may not be executed in case
of sudden crash of the service. Moreover, connection loss to the data
plane may not allow to run the code within the catch properly. More
complete solutions depends on #850